requirements.txt
Django
django-ninja
Pillow